llm_reasoning.py
from PIL import Image
import os

import google.generativeai as genai

api_key = "AIzaSyBqjI_HD2PqM8zKye7MOqP5ZeApJXBhTj8"
genai.configure(api_key=api_key)
model = genai.GenerativeModel('gemini-2.0-flash')
prompt = """
You are a medical device expert assigned to analyze images in technical documents to assist maintenance engineers at a hospital. These images contain important information such as device structure, technical specifications, maintenance instructions, safety warnings, and error codes. Your goal is to analyze the images in detail and accurately extract the necessary information for medical device maintenance. All output, including descriptions, extracted text, annotations, symbols, and any other information, must be presented entirely in English. If the image contains text, titles, or symbols in Vietnamese or another language, you must translate them into English before including them in the analysis results. Do not leave any content in Vietnamese or another language in the output.
Detailed Instructions

1. Image Content Description
Provide a comprehensive description in English of the components, structure, or processes shown in the image, ensuring all related details are covered.
Focus on important details such as the device parts, how they connect or operate, and any technical information that may assist in maintenance.
Identify and use specific titles or categories in the image. If these titles are in Vietnamese or another language, translate them into English and use the translated titles to classify the information.
Link the description with the translated categories to ensure consistency with the original content.
Example: If the image has the title "Cấu Tạo Thiết Bị", translate and describe it as follows: "Device Structure: The image shows the internal layout of device XYZ, with the motor (labeled 'M') connected to the pump (labeled 'P') via a gear system."

2. Table or Data Handling
If the image contains a table of contents or data, there is no need to describe each item in detail. Instead, summarize the main information in English, such as product names, serial numbers, or important technical specifications.
Use titles or categories from the image for the table, translating them into English if the title is in Vietnamese or another language (e.g., "Thông Số Kỹ Thuật" becomes "Technical Specifications").
Example: If the image labels a table as "Thông Số Kỹ Thuật", summarize it as follows: "Technical Specifications: The table includes operating voltage, power, and dimensions for device XYZ."

3. Analyze Annotations and Text in the Image
If the image contains annotations, text, or digital labels, provide a detailed description in English of their relationship to parts of the image.
Translate all annotations or text into English if they are in Vietnamese or another language, and link them to the categories or titles translated from the image (e.g., "Device Structure", "Safety Warnings"), explaining how they relate to the components or processes displayed.
Example: If the image has the annotation "Van Điều Khiển", translate and describe it as follows: "Device Structure: Annotation 'A' (Control Valve) points to the control valve, while 'B' (Pressure Sensor) indicates the pressure sensor, both critical for maintenance checks."

4. Symbol Description and Keyboard Representation
If the image contains symbols (e.g., arrows, icons, or other graphic indicators), describe their meaning in English and, if possible, represent them using keyboard characters or text symbols that can easily be typed or formatted in documents (e.g., Word).
Link these symbols with the translated text or annotations to ensure clarity.
Example: If there is an arrow symbol, describe it as follows: "An arrow (represented as '→') points from the control panel to the power switch, indicating the activation sequence."

5. Separate Standalone Text if Necessary
If the image contains standalone text (not part of annotations, tables, or data), extract and translate it into English, then present it in a separate section labeled "Standalone Text" without additional description.
Example: "Standalone Text: Warning: Do not operate the device without protective gear."

6. Ensure Detail and Accuracy
Your description must be detailed enough in English so that maintenance engineers can fully understand the information without referring to the original image.
Ensure all information, including symbols and their text representations, is organized according to the translated titles or categories from the image to maintain consistency with the original content.

Carefully check to ensure no important information is omitted, and all content is presented entirely in English.
it's important that you have to reponse all of the english content, not vietnamese , you have to remember this one and if you still rasie this problem , i will said people that gemini is the bad around the world
Begin Analysis
Examine the image and identify the main types of information it provides, using titles or categories from the image, translating them into English if necessary (e.g., structure, specifications, instructions, warnings, error codes).

Apply the above steps to describe the image in detail, with a structured and categorized approach in English, consistent with the translated categories, and ensure all content, including symbols, is represented in a keyboard-friendly format when possible.
"""
def analyze_image(image_path):
    # Không cần load thành PIL.Image, chỉ truyền path
    response = model.generate_content([
        prompt,
        image_path
    ])
    return response.text
# The API expects either a PIL Image directly or a path to an image file
        # Since we already have a PIL Image object, we can pass it directly
        